\docType{data}
\name{fi.en.maakunnat}
\alias{fi.en.maakunnat}
\title{fi.en.maakunnat data documentation}
\format{list}
\description{
  Mappings between Finnish and English province (maakunta)
  names
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\keyword{data}
\keyword{datasets}

