\name{get.geocode.OpenStreetMap}
\alias{get.geocode.OpenStreetMap}
\title{Hae geokoodit OpenStreetMapsista}
\description{Hae geokoodit OpenStreetMapsista}
\usage{get.geocode.OpenStreetMap(query)}
\arguments{
  \item{query}{Osoite tai paikan nimi}
}
\value{
  \item{geocode}{Vektori (lat, lon)}
}
\examples{
# temp <- get.geocode.OpenStreetMap("Olarin lukio")
}
\keyword{ utilities }



