\name{GetElectionResultsPresidentti2012}
\alias{GetElectionResultsPresidentti2012}
\title{Load presidential election 2012 results from HS Next}
\usage{
  GetElectionResultsPresidentti2012(election.round,
    level = NULL)
}
\arguments{
  \item{election.round}{Presidential election round (1/2)}

  \item{level}{Optional. Pick results for particular region
  type. Options: "municipalities" (kunnat)}
}
\value{
  Votes table
}
\description{
  Load presidential election 2012 results from HS Next
}
\author{
  Juuso Parkkinen and Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

