\name{GetHKK}
\alias{GetHKK}
\title{Retrieve HKK data}
\usage{
  GetHKK(which.data)
}
\arguments{
  \item{which.data}{A string. Specify the name of the HKK
  data set to retrieve. Currently available options:
  Aluejakokartat;Aanestysjakoalue;Seutukartta
  Rakennustietoruudukko; SeutuRAMAVA; key.KATAKER.}
}
\description{
  This script retrieves data from Helsinki Real Estate
  Department (Helsingin kaupunki kiinteistovirasto HKK)
  through the HKK website
  http://kartta.hel.fi/avoindata/index.html For details,
  see the HKK website
}
\details{
  The data copyright is on Helsingin kaupunkimittausosasto
  (C) 2011.
}
\author{
  Joona Lehtomäki
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

