\name{GetLukiot}
\alias{GetLukiot}
\title{Preprocess lukio data}
\usage{
  GetLukiot()
}
\value{
  list Data for all high schools and separately for the
  Helsinki Region
}
\description{
  Preprocess data about Finnish high school performance in
  year 2011
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

