\name{GetMaakuntainfo}
\alias{GetMaakuntainfo}
\title{Get Information about Finnish province}
\usage{
  GetMaakuntainfo()
}
\value{
  maakuntakartta data frame with map and info of the
  provinces
}
\description{
  Preprocess data about Finnish high school performance in
  year 2011
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

