\name{GetOikotie}
\alias{GetOikotie}
\title{Preprocess Oikotie data}
\usage{
  GetOikotie()
}
\value{
  list with Oikotie myynnit data
}
\description{
  Preprocess data about Finnish apartment prices from
  Oikotie in years 2010-2011
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

