\name{GetPresidentti2012Results}
\alias{GetPresidentti2012Results}
\title{Load presidential election 2012 results from HS Next}
\usage{
  GetPresidentti2012Results(election.round)
}
\arguments{
  \item{election.round}{Presidential election round (1/2)}
}
\value{
  Votes table
}
\description{
  Load presidential election 2012 results from HS Next
}
\author{
  Juuso Parkkinen and Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

