\name{GetPXTilastokeskus}
\alias{GetPXTilastokeskus}
\title{Get PC Axis data with custom preprocessing for PC Axis
files from Statistics Finland (Tilastokeskus) http://www.stat.fi/}
\usage{
  GetPXTilastokeskus(px)
}
\arguments{
  \item{px}{PC Axis object, or its URL.

  Returns:}
}
\value{
  PC Axis object.
}
\description{
  Arguments:
}
\examples{
# px <- GetPXTilastokeskus("http://pxweb2.stat.fi/database/StatFin/vrm/synt/080_synt_tau_203_fi.px")
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

