\name{GetWMSlayers}
\alias{GetWMSlayers}
\title{Get WMS layers}
\usage{
  GetWMSlayers(WMS)
}
\arguments{
  \item{WMS}{object}
}
\description{
  All WMS have a set of layers which correspond to the
  actual rasters. getLayers list all layers associated with
  a specific WSM object.
}
\note{
  NOT IMPLEMENTED YET.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.org}
}

