\name{hae.maakuntatiedot}
\alias{hae.maakuntatiedot}
\title{Maakuntien vaestontiheys}
\description{Maakuntien vaestontiheys}
\usage{hae.maakuntatiedot(url)}
\arguments{
  \item{url}{Wikipedian url, josta tiedot haetaan}
}
\value{
  \item{Maakunta}{Maakunta}
  \item{Pinta-ala}{Maakunnan pinta-ala}
  \item{Vakiluku}{Maakunnan vakiluku}
  \item{Vaestotiheys}{Maakunnan vaestotiheys}
}
\examples{
# tmp <- hae.maakuntatiedot(url)
}
\keyword{ utilities }



