\name{korvaa.skandit}
\alias{korvaa.skandit}
\title{Korvaa skandit}
\description{Korvaa skandit a/o:lla ongelmatilanteiden varalta.}
\usage{korvaa.skandit(s)}
\arguments{
  \item{s}{Merkkijono (character) tai merkkijonojen vektori.}
}
%\details{}
\value{Merkkijono, josta skandit on korvattu.}
\references{Viitetiedot, ks. citation("sorvi")}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
%\note{}
%\seealso{}
\examples{
# ss <- korvaa.skandit(s)
}
\keyword{ utilities }

