\name{kunta2maakunta}
\alias{kunta2maakunta}
\title{Etsi kunnalle vastaava maakunta.}
\description{}
\usage{kunta2maakunta(kunnat, map)}
\arguments{
  \item{kunnat}{Vektori kuntien nimia}
  \item{map}{data.frame josta loytyy tiedot kunnista ja maakunnista}
}
\value{Lista annettuja kuntia vastaavista maakunnista}
\examples{
cat("TBA\n")
}
\keyword{ utilities }

