\name{ListWMSurls}
\alias{ListWMSurls}
\title{List all the URLs associated with one of the standard WMSs.}
\usage{
  ListWMSurls()
}
\value{
  SpatialObject from the target WMS raster
}
\description{
  Standard WMSs are a group of well-known WMSs provided as
  part of soRvi. All URLs have a provider (organization
  etc.) and a WMS name and an URL.
}
\author{
  Joona Lehtomaki \email{joona.lehtomaki@gmail.org}
}
\seealso{
  buildServiceDesc listWMSurls
}

