\name{load.apurahat}
\alias{load.apurahat}
\title{Lataa apurahat data}
\description{Lataa apurahat-asuntosivuston hintatiedot asuntojen pyyntihinnoista ajalla 1/2010-8/2011. Data julkaistu HS Next -blogissa: http://www2.hs.fi/extrat/hsnext/apurahat-2005-2010.csv.}
\usage{load.apurahat()}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# load.apurahat()
}
\keyword{ load }



