\name{load.maakuntakartta}
\alias{load.maakuntakartta}
\title{Lataa maakuntakartta}
\description{Lataa Suomen kartta maakunnittain jaoteltuna. Karttadata haettu GADM database of Global Administrative Areas -tietokannasta: http://gadm.org/. Maakuntien asukasluvut haettu Tilastokeskukselta. }
\usage{load.maakuntakartta()}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# load.maakuntakartta()
}
\keyword{ load }



