\name{load.migration.data}
\alias{load.migration.data}
\title{Lataa vaeston muuttodata Maailmanpankin sivuilta}
\description{Lataa vaeston muuttodata Maailmanpankin sivuilta. Bilateral Migration and Remittances 2010.}
\usage{load.migration.data(countries)}
\arguments{
  \item{countries}{Haluttujen maiden nimet vektorina}
}
\value{
  \item{migration.dat}{Lista, jossa muuttovirta sisaan ja ulos halutusta maasta muihin maihin, seka muiden maiden nimet.}
}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# load.migration.data("Finland")
}
\keyword{ load }



