\docType{data}
\name{MML}
\alias{MML}
\title{MML data documentation}
\format{list}
\description{
  The MML data set contains preprocessed map data obtained
  from Maanmittauslaitos (MML; Finnish geographical
  agency). Copyright (C) MML 2011. The licenses allow
  modification and redistribution of the data. Preprocessed
  versions of most publicly available data sets are
  included. Certain shape data sets with preprocessing
  problems are currently ignored.
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org} is
  responsible of data preprocessing.
}
\references{
  \url{http://www.maanmittauslaitos.fi/aineistot-palvelut/digitaaliset-tuotteet/ilmaiset-aineistot/hankinta
  lisenssiehtojen mukaisesti:
  http://www.maanmittauslaitos.fi/node/6417}
}
\seealso{
  Description of the data identifiers:
  MML[["1_milj_Shape_etrs_shape"]]
  http://www.maanmittauslaitos.fi/digituotteet/yleiskartta-11-000-000;
  MML[["4_5_milj_shape_etrs-tm35fin"]]
  http://www.maanmittauslaitos.fi/digituotteet/yleiskartta-14-500-000
}
\keyword{data}
\keyword{datasets}

