\name{myynnit}
\alias{myynnit}
\docType{data}
\title{Oikotien myyntihintatiedot}
\description{Oikotie-asuntosivuston hintatiedot asuntojen pyyntihinnoista ajalla 1/2010-8/2011. Data julkaistu HS Next -blogissa: http://blogit.hs.fi/hsnext/asuntojen-vuokrat-ja-hintapyynnot-oikotiella-2010-2011. Lisenssi: Creative Commons Nimea-epakaupallinen-tarttuva 3.0.}
\usage{data(Oikotie)}
\format{list}
\keyword{misc}
