\docType{data}
\name{Oikotie}
\alias{Oikotie}
\title{Oikotie data documentation}
\format{data.frame}
\description{
  This data set contains information about apartment prices
  from Oikotie Helsinki Region apartment prices for years
  2010-2011 from Oikotie Contents: myynnit: National data,
  hr.myynnit: Helsinki redion data
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See cite(sorvi)
}
\keyword{data,}
\keyword{datasets}
\keyword{misc}

