\docType{data}
\name{PKS.aluejakokartat}
\alias{PKS.aluejakokartat}
\title{PKS.aluejakokartat data documentation}
\format{SpatialPolygonsDataFrame}
\description{
  This data set contains information about area boundaries
  for Helsinki Region Helsinki Region maps from Helsinki
  Region Infoshare Contents: pks.pienalue: Detailed area
  boundary data
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See cite(sorvi)
}
\keyword{data,}
\keyword{datasets}
\keyword{misc}

