\name{pks.df}
\alias{pks.df}
\docType{data}
\title{Paakaupunkiseudun aluejakokartat pienalueille}
\description{Paakaupunkiseudun aluejakokartat pienalueille. Sisaltaa Helsingin kaupunkimittausosaston aluejakoraja- aineistoa, 2011.}
\usage{data(PKS.aluejakokartat)}
\format{data.frame}
\details{}
\keyword{data}
