\docType{data}
\name{PKS.lukiot}
\alias{PKS.lukiot}
\title{PKS.lukiot data documentation}
\format{data.frame}
\description{
  This data set contains information about high schools in
  the Helsinki Region Helsinki Region high school
  performance data from year 2011 Contents: hr.lukiot:
  Helsinki Region high school data
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See cite(sorvi)
}
\keyword{data,}
\keyword{datasets}
\keyword{misc}

