\name{pks.pienalue}
\alias{pks.pienalue}
\docType{data}
\title{Paakaupunkiseudun aluejakokartat pienalueille}
\description{Paakaupunkiseudun aluejakokartat pienalueille. Sisaltaa Helsingin kaupunkimittausosaston aluejakoraja- aineistoa, 2011.}
\usage{data(PKS.aluejakokartat)}
\format{list}
\keyword{data}
