\name{preprocess.Oikotie}
\alias{preprocess.Oikotie}
\title{Esikasittele Oikotien myyntihintadataa}
\description{Esikasittele Oikotie-asuntosivuston hintatiedot asuntojen pyyntihinnoista ajalla 1/2010-8/2011. Data julkaistu HS Next -blogissa: http://blogit.hs.fi/hsnext/asuntojen-vuokrat-ja-hintapyynnot-oikotiella-2010-2011. Lisenssi: Creative Commons Nimea-epakaupallinen-tarttuva 3.0.}
\usage{preprocess.Oikotie()}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# preprocess.Oikotie()
}
\keyword{ preprocess }



