\name{preprocess.PKS.aluejakokartat}
\alias{preprocess.PKS.aluejakokartat}
\title{Preprocess 'aluejakokartat' from HRI}
\usage{
  preprocess.PKS.aluejakokartat()
}
\description{
  Preprocess 'aluejakokartat' from Helsinki Region
  Infoshare
}
\author{
  Juuso Parkkinen \email{juuso.parkkinen@gmail.org}
}

