\name{preprocess.PKS.lukiot}
\alias{preprocess.PKS.lukiot}
\title{Esikasittele lukiodataa}
\description{Esikasittele lukioiden YO-kirjoitusten pistekeskiarvodataa. Data haettu HS:n 31.5.2011 julkaisemasta artikkelista: http://www.hs.fi/kotimaa/artikkeli/Lukioiden+menestys+paremmuusjarjestyksessa+kevaan+2011+ylioppilaskirjoituksissa/1135266565425.}
\usage{preprocess.PKS.lukiot()}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# preprocess.PKS.lukiot()
}
\keyword{ preprocess }



