\name{preprocess.shape.mml}
\alias{preprocess.shape.mml}
\title{Retrieve MML data}
\usage{
  preprocess.shape.mml(sp)
}
\arguments{
  \item{sp}{Shape object (SpatialPolygonsDataFrame)}
}
\value{
  Shape object (from SpatialPolygonsDataFrame class)
}
\description{
  This script can be used for preprocessing of shape data
  from Finnish geographical agency (Maanmittauslaitos, MML)
  The data copyright is on (C) MML 2011.
}
\examples{
# sp <- preprocess.shape.mml(sp)
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

