\name{PreprocessPresidentti2012}
\alias{PreprocessPresidentti2012}
\title{Preprocess Presidentti2012 question data}
\usage{
  PreprocessPresidentti2012(questions)
}
\arguments{
  \item{questions}{output from questions <-
  GetPresidentti2012(category="questions", API=API)}
}
\value{
  list A list with the fields Questions and Choices
}
\description{
  Preprocess Presidentti2012 question data
}
\note{
  A wrapper
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

