\docType{package}
\name{sorvi-package}
\alias{sorvi}
\alias{sorvi-package}
\title{Algorithmic Tools for Open Data in Finland - soRvi: suomalainen avoimen datan tyokalupakki}
\description{
  Brief summary of the sorvi package
}
\details{
  \tabular{ll}{ Package: \tab sorvi\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2011\cr License: \tab FreeBSD\cr
  LazyLoad: \tab yes\cr }

  Algorithmic tools for Open Data in Finland -
  Laskennallisia valineita avointen tietoaineistojen
  hakuun, analysointiin ja visualisointiin
}
\examples{
library(sorvi)
}
\author{
  Leo Lahti, Juuso Parkkinen and Joona Lehtomaki
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
  \url{http://sorvi.r-forge.r-project.org}
}
\keyword{package}

