\name{Strip}
\alias{Strip}
\title{Strip string i.e. remove spaces from the beginning and end}
\usage{
  Strip(s)
}
\arguments{
  \item{s}{string or character vector}
}
\value{
  Stripped string
}
\description{
  Strip string i.e. remove spaces from the beginning and
  end
}
\examples{
#s2 <- Strip(s)
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

