\name{strstrip}
\alias{strstrip}
\title{Remove spaces from a string (single string or vector/list of strings).}
\usage{
  strstrip(s)
}
\arguments{
  \item{s}{string or vector/list of strings}
}
\value{
  string without spaces
}
\description{
  Remove spaces from a string (single string or vector/list
  of strings).
}
\examples{
# strstrip("a b") # returns "ab"
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

