\name{strstrip.single}
\alias{strstrip.single}
\title{Remove spaces from a single string}
\usage{
  strstrip.single(s)
}
\arguments{
  \item{s}{string}
}
\value{
  string without spaces
}
\description{
  Remove spaces from a single string
}
\examples{
# strstrip.single("a b") # returns "ab"
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See citation("sorvi")
}
\keyword{utilities}

