\name{fi.en.maakunnat}
\name{translations}
\alias{fi.en.maakunnat}
\alias{translations}
\docType{data}
\title{Kaannoksia}
\description{Nimien kaannoksia Suomen ja muiden kielten valilla.}
\usage{data(translations)}
\format{Lista}
\details{Maakuntien nimet suomi-englanti}
\examples{
# fi.en.maakunnat
}
\keyword{datasets}
