\name{get.geocode.GoogleMaps}
\alias{get.geocode.GoogleMaps}
\title{Hae geokoodit GoogleMapsista}
\description{Hae geokoodit GoogleMapsista}
\usage{get.geocode.GoogleMaps(str)}
\arguments{
  \item{str}{Tie}
}
\value{
  \item{geocode}{Vektori (lat, lon)}
}
\examples{
# temp <- get.geocode.GoogleMaps("Nallenpolku 2, 02110, FI")
}
\keyword{ utilities }



