\name{load.migration.data}
\alias{load.migration.data}
\title{Load migration data for given countries from Worldbank}
\usage{
  load.migration.data(countries)
}
\arguments{
  \item{countries}{a vector of country names}
}
\value{
  migration.dat list of migration data
}
\description{
  Data about countrywise migration in and out from given
  countries is obtained from the Worldbank database.
}
\author{
  Juuso Parkkinen
  \email{sorvi-commits@lists.r-forge.r-project.org}
}

