\name{MML}
\alias{MML}
\docType{data}
\title{Maanmittauslaitoksen kartta-aineistoja}
\description{Maanmittauslaitoksen shape-muotoisia kartta-aineistoja valmiiksi esikasitellyssa muodossa. (C) MML 2011.}
\details{Aineistot on haettu Maanmittauslaitoksen verkkosivuilta
http://www.maanmittauslaitos.fi/aineistot-palvelut/digitaaliset-tuotteet/ilmaiset-aineistot/hankinta lisenssiehtojen mukaisesti: http://www.maanmittauslaitos.fi/node/6417}
\usage{data(MML)}
\format{list}
\keyword{data}