\docType{data}
\name{MML}
\alias{MML}
\title{MML data documentation}
\format{list}
\description{
  The MML data set contains preprocessed map data obtained
  from Maanmittauslaitos (MML; Finnish geographical
  agency). Copyright (C) MML 2011. The licenses allow
  modification and redistribution of the data.
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org} is
  responsible of data preprocessing.
}
\references{
  \url{http://www.maanmittauslaitos.fi/aineistot-palvelut/digitaaliset-tuotteet/ilmaiset-aineistot/hankinta
  lisenssiehtojen mukaisesti:
  http://www.maanmittauslaitos.fi/node/6417}
}
\keyword{data}
\keyword{datasets}

