\name{pks.pienalue}
\alias{pks.pienalue}
\docType{data}
\title{PKS.aluejakokartat-data}
\description{PKS.aluejakokartat-data}
\usage{data(PKS.aluejakokartat)}
\format{list}
\details{Paakaupunkiseudun aluejakokartat. Sisaltaa Helsingin kaupunkimittausosaston aluejakoraja- aineistoa, 2011.}
\keyword{misc}
