\name{preprocess.PKS.aluejakokartat}
\alias{preprocess.PKS.aluejakokartat}
\title{Esikasittele PKS.aluejakokartat data}
\description{Esikasittele paakaupunkiseudun aluejakokartat dataa. Sisaltaa Helsingin kaupunkimittausosaston aluejakoraja- aineistoa, 2011.}
\usage{preprocess.PKS.lukiot()}
\examples{
# preprocess.PKS.aluejakokartat()
}
\keyword{ utilities }



