\name{preprocess.PKS.aluejakokartat}
\alias{preprocess.PKS.aluejakokartat}
\title{Esikasittele paakaupunkiseudun aluejakokarttoja}
\description{Esikasittele paakaupunkiseudun aluejakokartat dataa. Sisaltaa Helsingin kaupunkimittausosaston aluejakoraja- aineistoa, 2011. Data haettu taalta: http://www.hri.fi/fi/data/paakaupunkiseudun-aluejakokartat/.}
\usage{preprocess.PKS.lukiot()}
\author{Juuso Parkkinen \email{sorvi-commits@lists.r-forge.r-project.org}}
\examples{
# preprocess.PKS.aluejakokartat()
}
\keyword{ preprocess }



