\name{sorvi-package}
\alias{sorvi-package}
\alias{sorvi}
\docType{package}
\title{sorvi: avoimen datan tyokalupaketti}
\description{Laskennallisia valineita avoimilta rajapinnoilta
saatavien aineistojen hakuun, siivoamiseen, analysointiin ja
visualisointiin, painopisteena Suomea koskevat aiheet.}
\details{
\tabular{ll}{
Package: \tab sorvi\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2011-10-06\cr
License: \tab FreeBSD \cr
LazyLoad: \tab yes\cr
}
}
\author{Leo Lahti and Juuso Parkkinen.
Maintainer: Leo Lahti \email{leo.lahti@iki.fi} ja Juuso Parkkinen \email{juuso.parkkinen@gmail.com}}
\references{Leo Lahti ja Juuso Parkkinen: sorvi - avoimen datan
tyokalupaketti (2011). Ks. myos citation("sorvi").}
\keyword{package}
%\seealso{%%~~ \code{\link[somepkg:somepkg-package]{somepkg}} ~~}}

\examples{
  # Loading the package
  library(sorvi)
}
