\docType{data}
\name{translations}
\alias{translations}
\title{translations data documentation}
\format{list}
\description{
  This data set contains translations for common terms for
  which primary data is available in a non-Finnish language
  Contents: fi.en.maakunnat: Province names Finnish-English
}
\author{
  Leo Lahti
  \email{sorvi-commits@lists.r-forge.r-project.org}
}
\references{
  See cite(sorvi)
}
\keyword{data}
\keyword{datasets}
\keyword{misc}

